requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
paho-mqtt==2.1.0
pillow==11.0.0
ttkbootstrap==1.10.1
